feat : 하급기사 구현 완료

+++ Assets/Scripts/IInteractable.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public interface IInteractable
 {
-    public void TakeHit(int damage, UnitType unit, float stun);
+    public void TakeHit(float damage, UnitType unit, float stun);
 }
